feat(package): update version to 0.1.48 and refactor menu components to use new hooks
