+++ package/package.json
@@ -1,6 +1,6 @@
 {
   "name": "auera-ui",
-  "version": "0.1.47",
+  "version": "0.1.48",
   "main": "dist/auera-ui.cjs",
   "module": "dist/auera-ui.es.js",
   "types": "dist/index.d.ts",
