+++ package/src/context/menu.ts
@@ -1,22 +1,5 @@
-import { ModeType } from "@/types/auera-system";
+import { MenuContextProps } from "@/types/auera-context";
 import { createContext } from "react";
-
-type MenuContextProps = {
-  /**
-   * Indicates if the menu is currently open.
-   */
-  isOpen: boolean;
-  /**
-   * Function to open the menu.
-   */
-  onOpen: () => void;
-  /**
-   * Function to close the menu.
-   */
-  onClose: () => void;
-  mode: ModeType;
-  menuRef: React.RefObject<HTMLDivElement | null>;
-};
 
 export const MenuContext = createContext<MenuContextProps | undefined>(
   undefined
