+++ package/src/hook/useMenu.tsx
@@ -1,3 +1,4 @@
+import { MenuContext } from "@/context/menu";
 import { useContext } from "react";
 
 export const useMenu = () => {
