+++ package/src/hook/useStyleRules.ts
@@ -1,4 +1,3 @@
-import { i } from "motion/react-client";
 import {
   ApplyBy,
   ButtonVariant,
